## 3 September 2011
Version 2.3
* Widgets now support downloading of data. To enable this, you must add
 a custom attribute `allow_downloads=yes` and `dynamic=true` – see
 custom widget fields
* The widget `sub_text` is now displayed above the visual, and should
now be used to provide a title for the visual. The widget `title` should
be short. For example: a good widget title would be *Birth rates* and
then the sub_text should be *Birth rates for women aged 14-55 in Uganda.
Source: UBOS, 2015*.
* The More menu has been removed and now shows Data. It allows for downloading data.
The edit widget and view standalone have been moved to a separate menu, with a
spanner icon. The more info button has been moved to the Data menu, and will
also be displayed after the widget sub_text, so it is a good way to expand
on that text.
* Tiles: The tiles have been fixed! There were many bugs. Please start using
the tile template from the file system again (generic_widgets/w_tiles.html).
To ensure proper functioning, make sure you **either return** 3, 4 or 2 rows. Tiles
will be spanning the entire width now.
* Text widget: The text widget now supports proper HTML. The admin page has been
updated to support HTML as well. Use the 'text' field of the widget. Don't forget
to set a title too.
* 'widget.user_description' now supports HTML. This means you have more freedom
to provide a better description of what people are looking at.